+++ packages/unmock-xmlhttprequest/src/types.ts
@@ -1,0 +1,5 @@
+import * as crossFetch from "cross-fetch";
+
+export { Headers, Request, Response, fetch } from "cross-fetch";
+
+export { crossFetch };